pyproject.toml
[project]
name = "py_dct_txt"
version = "0.1.0"
description = "Python Boilerplate contains all the boilerplate you need to create a Python package."
readme = "README.md"
authors = [
  # {name = "???", email = "???"}
]
maintainers = [
  # {name = "???", email = "???"}
]
classifiers = [
    # TODO
]
license = {text = "MIT"}
dependencies = [
  "typer", "pyyaml",
]
requires-python = ">= 3.10"

[project.optional-dependencies]
test = [
    "coverage",  # testing
    "pytest",  # testing
    "ruff",  # linting
    "ty", # checking types
    "ipdb", # debugging
]

[project.urls]
bugs = "https://github.com/???/py_dct_txt/issues"
changelog = "https://github.com/???/py_dct_txt/blob/master/changelog.md"
homepage = "https://github.com/???/py_dct_txt"

[project.scripts]
py_dct_txt = "py_dct_txt.cli:app"

[tool.ty]
# All rules are enabled as "error" by default; no need to specify unless overriding.
# Example override: relax a rule for the entire project (uncomment if needed).
# rules.TY015 = "warn"  # For invalid-argument-type, warn instead of error.

[tool.ruff]
# line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.uv]
package = true
